Fixed up some inconsistencies between the Room and script schemas. Working now on the best way to have the filter buttons affect/observe the current CharacterProfile properties.

+++ app/src/main/java/dnd/jon/spellbook/SpellCreationActivity.java
@@ -235,7 +235,7 @@ public final class SpellCreationActivity extends AppCompatActivity {
         }
 
         // Set the checkboxes in the class selection grid
-        final List<Integer> spellClassIDs = repository.getClassIDs(spell);
+        final List<Long> spellClassIDs = repository.getClassIDs(spell);
         for (int i = 0; i < binding.classesSelectionGrid.getChildCount(); ++i) {
             final View view = binding.classesSelectionGrid.getChildAt(i);
             if (view instanceof RadioButton) {
